tests/submission/Ms2ger/loading-web-pages/the-window-object/window-prototype-chain.html

<!doctype html>
<meta charset=utf-8>
<title>Prototype chain of the window object</title>
<link rel="author" title="Ms2ger" href="ms2ger@gmail.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
function assert_class_name(obj, className) {
  assert_equals({}.toString.call(obj), "[object " + className + "]");
}
test(function() {
  assert_class_name(window, "Window");
}, "window object");
test(function() {
  var proto = Object.getPrototypeOf(window);
  assert_class_name(proto, "WindowPrototype");
  assert_equals(proto, Window.prototype);
}, "Window.prototype");
test(function() {
  var gsp = Object.getPrototypeOf(Object.getPrototypeOf(window));
  assert_class_name(gsp, "WindowProperties");
}, "Global scope polluter");
test(function() {
  var protoproto = Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(window)));
  assert_class_name(protoproto, "EventTargetPrototype");
  assert_equals(protoproto, EventTarget.prototype);
}, "EventTarget.prototype");
test(function() {
  var protoprotoproto = Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(Object.getPrototypeOf(window))));
  assert_class_name(protoprotoproto, "Object");
  assert_equals(protoprotoproto, Object.prototype);
}, "Object.prototype");
</script>
